requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9

# Frontend
streamlit==1.35.0
matplotlib==3.9.0
plotly==5.21.0

# OCR
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.3.0
easyocr==1.7.1
python-doctr==0.11.0

# Transformers & NLP
transformers==4.54.0
accelerate==0.29.3
sentencepiece==0.2.0
sentence-transformers==2.7.0
torch==2.3.0
scikit-learn==1.5.0
outlines==1.1.1

# CNN
torchvision==0.18.0

# Vector DB
chromadb==0.5.0

# DB and ORM
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Agents & LLM
langchain==0.2.1
openai==1.97.1

# Utilities
tqdm==4.66.4
regex==2024.5.15
python-dotenv==1.0.1
loguru==0.7.2

# Optional: Differential Privacy
opacus==1.4.0

# Extra
#postgresql
#datasets
